additional-libs feature for ledger-stellar dependencies (ledger signing) (#2011)
* Add stellar-ledger crate to additional-libs feature, and make optional

* Create a GenericLedger for when additional-libs feature is not enabled

* Change ledger::ledger fn to ledger::new

* Simplify ledger_impl

* Move stellar-ledger specific things to ledger_impl

* Move ledger feature switching to ledger.rs

* Enable additional-libs in emulator workflow
